Extract separate JsonUtils class containing all JSON related parsing/formatting previously found in ProtobufUtil
* this will ease future maintenance
